+++ AddConfig.axaml
@@ -1,0 +1,15 @@
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="200" d:DesignHeight="80" Width="200" Height="80" CanResize="False" 
+        x:Class="ws.AddConfig"
+        Title="Add Config...">
+    <Grid ColumnDefinitions="Auto, *" RowDefinitions="Auto,Auto" VerticalAlignment="Center" Margin="10,10,10,10">
+        
+        <Label VerticalAlignment="Center" VerticalContentAlignment="Center" >Header: </Label>
+        <TextBox Grid.Column="1" x:Name="Header"></TextBox>
+        <Button Grid.Row="1" Grid.ColumnSpan="2" Margin="0,10,0,0" Click="Button_OnClick">Ok</Button>
+
+    </Grid>
+</Window>
