+++ MainView.axaml
@@ -1,0 +1,41 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="ws.MainView">
+    <DockPanel>
+		
+				<Grid Margin="0,5,0,0">
+					<Grid.RowDefinitions>
+						<RowDefinition Height="200"/>
+						<RowDefinition/>
+					</Grid.RowDefinitions>
+					<StackPanel>
+						<Grid ColumnDefinitions="*,*" >
+							<TextBox Watermark="Host" x:Name="host" TextChanged="Host_OnTextChanged" VerticalContentAlignment="Center"></TextBox>
+							<NumericUpDown Watermark="Port" x:Name="port" AllowSpin="True" Maximum="65535" Value="" FormatString="0" Minimum="1" Grid.Column="1" ></NumericUpDown>
+						</Grid>
+						<Grid ColumnDefinitions="*,*">
+							<Button Margin="0 10 0 0 " FontSize="9" VerticalContentAlignment="Center" Click="Button_Click" x:Name="connectbutton">Connect in TCP</Button>
+
+							<Button Margin="0 10 0 0 " FontSize="9" Grid.Column="1" VerticalContentAlignment="Center" Click="tlsConnect" x:Name="connecttls"   >Connect in TCP with TLS</Button>
+						</Grid>
+						<Button Margin=" 0 10 0 0" IsEnabled="False" Click="Button_Click_2" x:Name="disconnect">Disconnect</Button>
+						<TextBox PasswordChar="*" Watermark="Password"   x:Name="password" Margin="0 10 0 0" ></TextBox>
+						<Button Margin=" 0 10 0 0 " Click="authorizeButton">Authorize</Button>
+						<CheckBox Content="Text wrap" Margin="0 5 0 0" Click="changeWrap" x:Name="checkbox"/>
+					</StackPanel>
+					<DockPanel Grid.Row="1" VerticalAlignment="Stretch" >
+						<TextBox DockPanel.Dock="Top" Watermark="Commands" x:Name="command" IsEnabled="False"></TextBox>
+						<Button DockPanel.Dock="Top" Click="Button_Click_1" IsEnabled="False" x:Name="sendbutton">Send</Button>
+						<ScrollViewer x:Name="scrollv">
+							<TextBox DockPanel.Dock="Bottom"  IsReadOnly="True"  VerticalContentAlignment="Stretch" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" x:Name="output"  />
+	
+						</ScrollViewer>
+				  
+					</DockPanel>
+				</Grid>  
+			
+	</DockPanel>
+</UserControl>
