+++ MainWindow.axaml
@@ -4,54 +4,19 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="300" d:DesignHeight="500" Width="300" Height="500"
         x:Class="ws.MainWindow"
-        Title="IWTCMS" Padding="0,0,0,0" MinHeight="500" MinWidth="300" Icon="icon.ico">
-	<DockPanel>
-		<Menu DockPanel.Dock="Top">
-			<MenuItem Header="_File">
-				<MenuItem Header="_New Config"/>
-			</MenuItem>
-			<MenuItem Header="_Help">
-				<MenuItem Header="_About" Click="ShowAbout"></MenuItem>
-			</MenuItem>
-		</Menu>
-		<TabControl x:Name="TabControl" Padding="5,0,5,5" >
-		
-			<TabItem Header="Default">
-				<Grid >
-					<Grid.RowDefinitions>
-						<RowDefinition Height="200"/>
-						<RowDefinition/>
-					</Grid.RowDefinitions>
-					<StackPanel>
-						<Grid ColumnDefinitions="*,*" >
-							<TextBox Watermark="Host" x:Name="host" VerticalContentAlignment="Center">192.168.1.166</TextBox>
-							<NumericUpDown Watermark="Port" x:Name="port" AllowSpin="True" Maximum="65535" Value="25566" FormatString="0" Minimum="1" Grid.Column="1" ></NumericUpDown>
-						</Grid>
-						<Grid ColumnDefinitions="*,*">
-							<Button Margin="0 10 0 0 " FontSize="9" VerticalContentAlignment="Center" Click="Button_Click" x:Name="connectbutton">Connect in TCP</Button>
+        Title="IWTCMS" Padding="0,0,0,0" MinHeight="500" MinWidth="300" Icon="icon.webp">
+    <DockPanel>
+        <Menu DockPanel.Dock="Top">
+            <Menu.Styles></Menu.Styles>
+            <MenuItem Header="_File">
+                <MenuItem Header="_New Config" Click="AddConfigMenuItemOnClick" />
+            </MenuItem>
+            <MenuItem Header="_Help">
+                <MenuItem Header="_About" Click="ShowAbout"></MenuItem>
+            </MenuItem>
+        </Menu>
+        <TabControl x:Name="TabControl" Padding="5,0,5,5" />
+    </DockPanel>
 
-							<Button Margin="0 10 0 0 " FontSize="9" Grid.Column="1" VerticalContentAlignment="Center" Click="tlsConnect" x:Name="connecttls"   >Connect in TCP with TLS</Button>
-						</Grid>
-						<Button Margin=" 0 10 0 0" IsEnabled="False" Click="Button_Click_2" x:Name="disconnect">Disconnect</Button>
-						<TextBox PasswordChar="*" Watermark="Password"   x:Name="password" Margin="0 10 0 0" ></TextBox>
-						<Button Margin=" 0 10 0 0 " Click="authorizeButton">Authorize</Button>
-						<CheckBox Content="Text wrap" Margin="0 5 0 0" Click="changeWrap" x:Name="checkbox"/>
-					</StackPanel>
-					<DockPanel Grid.Row="1" VerticalAlignment="Stretch" >
-						<TextBox DockPanel.Dock="Top" Watermark="Commands" x:Name="command" IsEnabled="False"></TextBox>
-						<Button DockPanel.Dock="Top" Click="Button_Click_1" IsEnabled="False" x:Name="sendbutton">Send</Button>
-						<ScrollViewer x:Name="scrollv">
-							<TextBox DockPanel.Dock="Bottom"  IsReadOnly="True"  VerticalContentAlignment="Stretch" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" x:Name="output"  />
-	
-						</ScrollViewer>
-				  
-					</DockPanel>
-				</Grid>  
-			</TabItem>
-		
-		</TabControl>
-	</DockPanel>
-	
-	
 
 </Window>
