+++ ws.csproj
@@ -7,7 +7,7 @@
     <ApplicationManifest>app.manifest</ApplicationManifest>
     <AvaloniaUseCompiledBindingsByDefault>true</AvaloniaUseCompiledBindingsByDefault>
 	  <EnableCompressionInSingleFile>true</EnableCompressionInSingleFile>
-	  <PublishTrimmed>true</PublishTrimmed>
+	  <PublishTrimmed>false</PublishTrimmed>
 	  <IncludeNativeLibrariesForSelfExtract>true</IncludeNativeLibrariesForSelfExtract>
 	  <IncludeAllContentForSelfExtract>true</IncludeAllContentForSelfExtract>
     <PublishSingleFile>true</PublishSingleFile>
@@ -20,9 +20,7 @@
   <ItemGroup>
     <EmbeddedResource Include="cert.pfx" />
     <None Remove="icon.webp" />
-    <EmbeddedResource Include="icon.webp">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </EmbeddedResource>
+    <AvaloniaResource Include="icon.webp" />
   </ItemGroup>
   <ItemGroup>
     <AvaloniaResource Include="icon.ico" />
